# Improve the performance of column `scatter`