<!---
A technical comment, you are free to remove or leave it as it is when PR is created
The following categories are used in the next scripts, update them accordingly
utils/changelog/changelog.py
tests/ci/run_check.py
-->
### Changelog category (leave one):
- Performance Improvement



### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
...
Improve the performance of column `scatter` by writing into columns sequentially, without switching destination columns frequently which reduce cache miss.

we compare with the previous implementation as following.

query
```sql
-- column scatter is used in parallel_hash
SELECT
    t1.msgtype,
    count(*)
FROM
(
    SELECT
        rip,
        rtime,
        msgid,
        msgtype
    FROM test_data
    WHERE (day = '2023-03-06') AND (hour = '00')
) AS t1
ANY INNER JOIN
(
    SELECT
        rip,
        rtime,
        msgid,
        msgtype
    FROM test_data
    WHERE (day = '2023-03-06') AND (hour = '00')
) AS t2 ON (t1.rip = t2.rip) AND (t1.msgid = t2.msgid) AND (t1.rtime = t2.rtime) AND (t1.msgtype = t2.msgtype)
GROUP BY t1.msgtype
SETTINGS max_threads = 32, join_algorithm = 'parallel_hash'
```

result

|  threads   | before  | after |
|  ----  | ----  | ----  |
| 1 | 6 rows in set. Elapsed: 46.652 sec. Processed 37.40 million rows, 4.53 GB (801.77 thousand rows/s., 97.12 MB/s.) | 6 rows in set. Elapsed: 45.250 sec. Processed 37.40 million rows, 4.53 GB (826.61 thousand rows/s., 100.13 MB/s.) |
| 2  | 6 rows in set. Elapsed: 30.992 sec. Processed 37.40 million rows, 4.53 GB (1.21 million rows/s., 146.19 MB/s.) | 6 rows in set. Elapsed: 26.425 sec. Processed 37.40 million rows, 4.53 GB (1.42 million rows/s., 171.46 MB/s.) |
| 4 | 6 rows in set. Elapsed: 21.150 sec. Processed 37.40 million rows, 4.53 GB (1.77 million rows/s., 214.23 MB/s.) | 6 rows in set. Elapsed: 18.612 sec. Processed 37.40 million rows, 4.53 GB (2.01 million rows/s., 243.43 MB/s.) |
| 8  | 6 rows in set. Elapsed: 18.219 sec. Processed 37.40 million rows, 4.53 GB (2.05 million rows/s., 248.69 MB/s.) | 6 rows in set. Elapsed: 16.080 sec. Processed 37.40 million rows, 4.53 GB (2.33 million rows/s., 281.76 MB/s.) |
| 16 | 6 rows in set. Elapsed: 11.162 sec. Processed 37.40 million rows, 4.53 GB (3.35 million rows/s., 405.92 MB/s.) | 6 rows in set. Elapsed: 11.620 sec. Processed 37.40 million rows, 4.53 GB (3.22 million rows/s., 389.92 MB/s.) |
| 32  | 6 rows in set. Elapsed: 12.200 sec. Processed 37.40 million rows, 4.53 GB (3.07 million rows/s., 371.39 MB/s.) | 6 rows in set. Elapsed: 11.879 sec. Processed 37.40 million rows, 4.53 GB (3.15 million rows/s., 381.41 MB/s.) |

### Documentation entry for user-facing changes

- [ ] Documentation is written (mandatory for new features)

<!---
Directly edit documentation source files in the "docs" folder with the same pull-request as code changes

or

Add a user-readable short description of the changes that should be added to docs.clickhouse.com below.

At a minimum, the following information should be added (but add more as needed).
- Motivation: Why is this function, table engine, etc. useful to ClickHouse users?

- Parameters: If the feature being added takes arguments, options or is influenced by settings, please list them below with a brief explanation.

- Example use: A query or command.
-->


> Information about CI checks: https://clickhouse.com/docs/en/development/continuous-integration/