@vdimir 2023-03-10
No change for `parallel_hash_join` in https://s3.amazonaws.com/clickhouse-test-reports/47329/8aa18ca1406abc106de682268193d453266866f2/performance_comparison_[2/4]/report.html 🤔 

And could you please check slow downs reported in the reports, can they be related?