@lgbo-ustc 2023-03-14
With small source size, this implementation will be better.
Maybe we try it later.